+++ src/main/java/roomescape/business/service/ReservationService.java
@@ -9,10 +9,10 @@ import org.springframework.stereotype.Service;
 import roomescape.business.Reservation;
 import roomescape.business.ReservationTheme;
 import roomescape.business.ReservationTime;
-import roomescape.business.dto.ReservationRequestDto;
-import roomescape.business.dto.ReservationResponseDto;
-import roomescape.business.dto.ReservationThemeResponseDto;
-import roomescape.business.dto.ReservationTimeResponseDto;
+import roomescape.presentation.dto.ReservationRequestDto;
+import roomescape.presentation.dto.ReservationResponseDto;
+import roomescape.presentation.dto.ReservationThemeResponseDto;
+import roomescape.presentation.dto.ReservationTimeResponseDto;
 import roomescape.persistence.ReservationRepository;
 import roomescape.persistence.ReservationThemeRepository;
 import roomescape.persistence.ReservationTimeRepository;
