+++ src/main/java/roomescape/persistence/H2ReservationTimeRepository.java
@@ -10,7 +10,7 @@ import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.jdbc.core.simple.SimpleJdbcInsert;
 import org.springframework.stereotype.Repository;
 import roomescape.business.ReservationTime;
-import roomescape.business.dto.AvailableTimesResponseDto;
+import roomescape.presentation.dto.AvailableTimesResponseDto;
 
 @Repository
 public class H2ReservationTimeRepository implements ReservationTimeRepository {
