+++ src/main/java/roomescape/presentation/dto/ReservationRequestDto.java
@@ -1,4 +1,4 @@
-package roomescape.business.dto;
+package roomescape.presentation.dto;
 
 import jakarta.validation.constraints.NotEmpty;
 import jakarta.validation.constraints.NotNull;
