+++ src/main/java/roomescape/presentation/dto/ReservationResponseDto.java
@@ -1,4 +1,4 @@
-package roomescape.business.dto;
+package roomescape.presentation.dto;
 
 import java.time.LocalDate;
