+++ src/main/java/roomescape/presentation/dto/ReservationThemeResponseDto.java
@@ -1,4 +1,4 @@
-package roomescape.business.dto;
+package roomescape.presentation.dto;
 
 public record ReservationThemeResponseDto(
         long id,
