+++ src/main/java/roomescape/presentation/dto/ReservationTimeRequestDto.java
@@ -1,4 +1,4 @@
-package roomescape.business.dto;
+package roomescape.presentation.dto;
 
 import jakarta.validation.constraints.NotNull;
 import java.time.LocalTime;
