+++ src/main/java/roomescape/presentation/dto/ReservationTimeResponseDto.java
@@ -1,4 +1,4 @@
-package roomescape.business.dto;
+package roomescape.presentation.dto;
 
 import java.time.LocalTime;
