+++ src/test/java/roomescape/business/service/ReservationTimeServiceTest.java
@@ -13,8 +13,8 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import roomescape.business.Reservation;
 import roomescape.business.ReservationTime;
-import roomescape.business.dto.ReservationTimeRequestDto;
-import roomescape.business.dto.ReservationTimeResponseDto;
+import roomescape.presentation.dto.ReservationTimeRequestDto;
+import roomescape.presentation.dto.ReservationTimeResponseDto;
 import roomescape.business.fakerepository.FakeReservationRepository;
 import roomescape.business.fakerepository.FakeReservationTimeRepository;
 import roomescape.persistence.ReservationRepository;
